modules/location_consumer/requirements.txt
kafka-python==2.0.2
SQLAlchemy==1.4.45
psycopg2-binary
